+++ inst/shiny-examples/BCGcalc/server.R
@@ -20,7 +20,7 @@ shinyServer(function(input, output) {
       return("..No file uploaded yet...")
     }##IF~is.null~END
     
-    return(paste0("'",inFile$name,"'"))
+    return(paste0("'", inFile$name, "'"))
     
   })## fn_input_display_bcg
   
@@ -31,7 +31,7 @@ shinyServer(function(input, output) {
       return("..No file uploaded yet...")
     }##IF~is.null~END
     
-    return(paste0("'",inFile$name,"'"))
+    return(paste0("'", inFile$name, "'"))
     
   })## fn_input_display_taxatrans
   
@@ -42,7 +42,7 @@ shinyServer(function(input, output) {
       return("..No file uploaded yet...")
     }##IF~is.null~END
     
-    return(paste0("'",inFile$name,"'"))
+    return(paste0("'", inFile$name, "'"))
     
   })## fn_input_display_indexclass
   
@@ -53,9 +53,42 @@ shinyServer(function(input, output) {
       return("..No file uploaded yet...")
     }##IF~is.null~END
     
-    return(paste0("'",inFile$name,"'"))
+    return(paste0("'", inFile$name, "'"))
     
-  })## fn_input_display_indexclass
+  })## fn_input_display_met_therm
+  
+  output$fn_input_display_modtherm <- renderText({
+    inFile <- input$fn_input
+    
+    if (is.null(inFile)) {
+      return("..No file uploaded yet...")
+    }##IF~is.null~END
+    
+    return(paste0("'", inFile$name, "'"))
+    
+  })## fn_input_display_modtherm
+  
+  output$fn_input_display_mtti <- renderText({
+    inFile <- input$fn_input
+    
+    if (is.null(inFile)) {
+      return("..No file uploaded yet...")
+    }##IF~is.null~END
+    
+    return(paste0("'", inFile$name, "'"))
+    
+  })## fn_input_display_mtti
+  
+  output$fn_input_display_bdi <- renderText({
+    inFile <- input$fn_input
+    
+    if (is.null(inFile)) {
+      return("..No file uploaded yet...")
+    }##IF~is.null~END
+    
+    return(paste0("'", inFile$name, "'"))
+    
+  })## fn_input_display_bdi
   
   # IMPORT ----
   file_watch <- reactive({
@@ -164,6 +197,9 @@ shinyServer(function(input, output) {
     shinyjs::enable("b_taxatrans_calc")
     shinyjs::enable("b_indexclass_calc")
     shinyjs::enable("b_calc_met_therm")
+    shinyjs::enable("b_calc_modtherm")
+    # shinyjs::enable("b_calc_mtti")
+    # shinyjs::enable("b_calc_bdi")
     
     # update cb_taxatrans_sum 
     # doesn't work here as timing is after the file is created
@@ -1931,6 +1967,345 @@ shinyServer(function(input, output) {
   )##download ~ MergeFiles
   
   # FUZZY THERMAL ----
+  # b_Calc_modtherm ----
+  observeEvent(input$b_calc_modtherm, {
+    shiny::withProgress({
+      
+      ## Calc, 0, Set Up Shiny Code ----
+      
+      prog_detail <- "Calculation, Thermal Model..."
+      message(paste0("\n", prog_detail))
+      
+      # Number of increments
+      prog_n <- 10
+      prog_sleep <- 0.25
+      
+      # Calc, 1, Initialize ----
+      prog_detail <- "Initialize Data"
+      message(paste0("\n", prog_detail))
+      # Increment the progress bar, and update the detail text.
+      incProgress(1/prog_n, detail = prog_detail)
+      Sys.sleep(prog_sleep)
+      
+      # button, disable, download
+      shinyjs::disable("b_bcg_download")
+      
+      # data
+      inFile <- input$fn_input
+      fn_input_base <- tools::file_path_sans_ext(inFile$name)
+      message(paste0("Import, file name, base: ", fn_input_base))
+      df_input <- read.delim(inFile$datapath
+                             , header = TRUE
+                             , sep = input$sep
+                             , stringsAsFactors = FALSE)
+      # QC, FAIL if TRUE
+      if (is.null(df_input)) {
+        return(NULL)
+      }
+      
+      # QC, names to upper case
+      names(df_input) <- toupper(names(df_input))
+      
+      # Calc, 2, Exclude Taxa ----
+      prog_detail <- "Calculate, Exclude Taxa"
+      message(paste0("\n", prog_detail))
+      # Increment the progress bar, and update the detail text.
+      incProgress(1/prog_n, detail = prog_detail)
+      Sys.sleep(prog_sleep)
+      # Calc
+      
+      message(paste0("User response to generate ExclTaxa = ", input$ExclTaxa))
+      
+      if (input$ExclTaxa) {
+        ## Get TaxaLevel names present in user file
+        phylo_all <- c("Kingdom"
+                       , "Phylum"
+                       , "SubPhylum"
+                       , "Class"
+                       , "SubClass"
+                       , "Order"
+                       , "SubOrder"
+                       , "InfraOrder"
+                       , "SuperFamily"
+                       , "Family"
+                       , "SubFamily"
+                       , "Tribe"
+                       , "Genus"
+                       , "SubGenus"
+                       , "Species"
+                       , "Variety")
+        phylo_all <- toupper(phylo_all) # so matches rest of file
+        
+        # case and matching of taxa levels handled inside of markExluded 
+        
+        # overwrite current data frame
+        df_input <- BioMonTools::markExcluded(df_samptax = df_input
+                                              , SampID = "SAMPLEID"
+                                              , TaxaID = "TAXAID"
+                                              , TaxaCount = "N_TAXA"
+                                              , Exclude = "EXCLUDE"
+                                              , TaxaLevels = phylo_all
+                                              , Exceptions = NA)
+        
+        # Save Results
+        fn_excl <- paste0(fn_input_base, "_modtherm_1markexcl.csv")
+        dn_excl <- path_results
+        pn_excl <- file.path(dn_excl, fn_excl)
+        write.csv(df_input, pn_excl, row.names = FALSE)
+        
+      }## IF ~ input$ExclTaxa
+      
+      
+      # Calc, 3, BCG Flag Cols ----
+      # get columns from Flags (non-metrics) to carry through
+      prog_detail <- "Calculate, Keep BCG Model Columns"
+      message(paste0("\n", prog_detail))
+      # Increment the progress bar, and update the detail text.
+      incProgress(1/prog_n, detail = prog_detail)
+      Sys.sleep(prog_sleep)
+      # Rules - should all be metrics but leaving here just in case
+      # Flags - not always metrics,
+      # Index Name for import data
+      import_IndexName <- unique(df_input$INDEX_NAME)
+      # QC Flags for chosen BCG model (non-metrics)
+      cols_flags <- unique(df_checks[df_checks$Index_Name == import_IndexName
+                                     , "Metric_Name"])
+      # can also add other columns to keep if feel so inclined
+      cols_flags_keep <- cols_flags[cols_flags %in% names(df_input)]
+      
+      
+      # Calc, 3b, Rules ----
+      prog_detail <- "Calculate, BCG Rules"
+      message(paste0("\n", prog_detail))
+      message(paste0("Community = ", input$si_community))
+      # Increment the progress bar, and update the detail text.
+      incProgress(1/prog_n, detail = prog_detail)
+      Sys.sleep(prog_sleep)
+      # filter for data Index_Name in data (drop 2 extra columns)
+      df_rules <- df_bcg_models[df_bcg_models$Index_Name == import_IndexName
+                                , !names(df_bcg_models) %in% c("SITE_TYPE", "INDEX_REGION")]
+      # Save
+      fn_rules <- paste0(fn_input_base, "_modtherm_3metrules.csv")
+      dn_rules <- path_results
+      pn_rules <- file.path(dn_rules, fn_rules)
+      write.csv(df_rules, pn_rules, row.names = FALSE)
+      
+      # Calc, 4, MetVal----
+      prog_detail <- "Calculate, Metric, Values"
+      message(paste0("\n", prog_detail))
+      message(paste0("Community = ", input$si_community))
+      # Increment the progress bar, and update the detail text.
+      incProgress(1/prog_n, detail = prog_detail)
+      Sys.sleep(prog_sleep)
+      # Calc
+      # QC
+      # df_input <- read.csv(file.path("inst", "extdata", "Data_BCG_PacNW.csv"))
+      # df_metval <- BioMonTools::metric.values(df_input, "bugs", boo.Shiny = TRUE)
+      
+      if (length(cols_flags_keep) > 0) {
+        # keep extra cols from Flags (non-metric)
+        df_metval <- BioMonTools::metric.values(df_input
+                                                , input$si_community
+                                                , fun.cols2keep = cols_flags_keep
+                                                , boo.Shiny = TRUE
+                                                , verbose = TRUE)
+      } else {
+        df_metval <- BioMonTools::metric.values(df_input
+                                                , input$si_community
+                                                , boo.Shiny = TRUE
+                                                , verbose = TRUE)
+      }## IF ~ length(col_rules_keep)
+      
+      #df_metval$INDEX_CLASS <- df_metval$INDEX_CLASS
+      ## Save Results ----
+      fn_metval <- paste0(fn_input_base, "_modtherm_2metval_all.csv")
+      dn_metval <- path_results
+      pn_metval <- file.path(dn_metval, fn_metval)
+      write.csv(df_metval, pn_metval, row.names = FALSE)
+      
+      ## Save Results (BCG) ----
+      # Munge
+      ## Model and QC Flag metrics only
+      # cols_flags defined above
+      cols_model_metrics <- unique(df_bcg_models[
+        df_bcg_models$Index_Name == import_IndexName, "Metric_Name"])
+      cols_req <- c("SAMPLEID", "INDEX_NAME", "INDEX_CLASS"
+                    , "ni_total", "nt_total")
+      cols_metrics_flags_keep <- unique(c(cols_req
+                                          , cols_flags
+                                          , cols_model_metrics))
+      df_metval_slim <- df_metval[, names(df_metval) %in% cols_metrics_flags_keep]
+      # Save
+      fn_metval_slim <- paste0(fn_input_base, "_modtherm_2metval_BCG.csv")
+      dn_metval_slim <- path_results
+      pn_metval_slim <- file.path(dn_metval_slim, fn_metval_slim)
+      write.csv(df_metval_slim, pn_metval_slim, row.names = FALSE)
+      
+      
+      # Calc, 5, MetMemb----
+      prog_detail <- "Calculate, Metric, Membership"
+      message(paste0("\n", prog_detail))
+      # Increment the progress bar, and update the detail text.
+      incProgress(1/prog_n, detail = prog_detail)
+      Sys.sleep(prog_sleep)
+      # Calc
+      df_metmemb <- BCGcalc::BCG.Metric.Membership(df_metval, df_bcg_models)
+      # Save Results
+      fn_metmemb <- paste0(fn_input_base, "_modtherm_3metmemb.csv")
+      dn_metmemb <- path_results
+      pn_metmemb <- file.path(dn_metmemb, fn_metmemb)
+      write.csv(df_metmemb, pn_metmemb, row.names = FALSE)
+      
+      
+      # Calc, 6, LevMemb----
+      prog_detail <- "Calculate, Level, Membership"
+      message(paste0("\n", prog_detail))
+      # Increment the progress bar, and update the detail text.
+      incProgress(1/prog_n, detail = prog_detail)
+      Sys.sleep(prog_sleep)
+      # Calc
+      df_levmemb <- BCGcalc::BCG.Level.Membership(df_metmemb, df_bcg_models)
+      # Save Results
+      fn_levmemb <- paste0(fn_input_base, "_modtherm_4levmemb.csv")
+      dn_levmemb <- path_results
+      pn_levmemb <- file.path(dn_levmemb, fn_levmemb)
+      write.csv(df_levmemb, pn_levmemb, row.names = FALSE)
+      
+      
+      # Calc, 7, LevAssign----
+      prog_detail <- "Calculate, Level, Assignment"
+      message(paste0("\n", prog_detail))
+      # Increment the progress bar, and update the detail text.
+      incProgress(1/prog_n, detail = prog_detail)
+      Sys.sleep(prog_sleep)
+      # Calc
+      df_levassign <- BCGcalc::BCG.Level.Assignment(df_levmemb)
+      # Save Results
+      fn_levassign <- paste0(fn_input_base, "_modtherm_5levassign.csv")
+      dn_levassign <- path_results
+      pn_levassign <- file.path(dn_levassign, fn_levassign)
+      write.csv(df_levassign, pn_levassign, row.names = FALSE)
+      
+      
+      # Calc, 8, QC Flags----
+      prog_detail <- "Calculate, QC Flags"
+      message(paste0("\n", prog_detail))
+      # Increment the progress bar, and update the detail text.
+      incProgress(1/prog_n, detail = prog_detail)
+      Sys.sleep(prog_sleep)
+      
+      # Calc
+      # df_checks loaded in global.R
+      df_flags <- BioMonTools::qc.checks(df_metval, df_checks)
+      # Change terminology; PASS/FAIL to NA/flag
+      df_flags[, "FLAG"][df_flags[, "FLAG"] == "FAIL"] <- "flag"
+      df_flags[, "FLAG"][df_flags[, "FLAG"] == "PASS"] <- NA
+      # long to wide format
+      df_flags_wide <- reshape2::dcast(df_flags
+                                       , SAMPLEID ~ CHECKNAME
+                                       , value.var = "FLAG")
+      # Calc number of "flag"s by row.
+      df_flags_wide$NumFlags <- rowSums(df_flags_wide == "flag", na.rm = TRUE)
+      # Rearrange columns
+      NumCols <- ncol(df_flags_wide)
+      df_flags_wide <- df_flags_wide[, c(1, NumCols, 2:(NumCols - 1))]
+      # Merge Levels and Flags
+      df_lev_flags <- merge(df_levassign
+                            , df_flags_wide
+                            , by.x = "SampleID"
+                            , by.y = "SAMPLEID"
+                            , all.x = TRUE)
+      # Summarize Flags
+      df_lev_flags_summ <- as.data.frame.matrix(table(df_flags[, "CHECKNAME"]
+                                                      , df_flags[, "FLAG"]
+                                                      , useNA = "ifany"))
+      
+      # Save Flags Summary
+      fn_levflags <- paste0(fn_input_base, "_modtherm_6levflags.csv")
+      dn_levflags <- path_results
+      pn_levflags <- file.path(dn_levflags, fn_levflags)
+      write.csv(df_lev_flags_summ, pn_levflags, row.names = TRUE)
+      
+      # Create Results
+      df_results <- df_lev_flags[, !names(df_lev_flags) %in% c(paste0("L", 1:6))]
+      ## remove L1:6
+      
+      # Save Results
+      fn_results <- paste0(fn_input_base, "_modtherm_RESULTS.csv")
+      dn_results <- path_results
+      pn_results <- file.path(dn_results, fn_results)
+      write.csv(df_results, pn_results, row.names = FALSE)
+      
+      
+      # Calc, 8b, QC Flag Metrics ----
+      # create
+      col2keep <- c("SAMPLEID", "INDEX_NAME", "INDEX_CLASS", "METRIC_NAME"
+                    , "CHECKNAME", "METRIC_VALUE", "SYMBOL", "VALUE", "FLAG")
+      df_metflags <- df_flags[, col2keep]
+      # save
+      fn_metflags <- paste0(fn_input_base, "_modtherm_6metflags.csv")
+      dn_metflags <- path_results
+      pn_metflags <- file.path(dn_metflags, fn_metflags)
+      write.csv(df_metflags, pn_metflags, row.names = FALSE)
+      
+      
+      # Calc, 9, RMD----
+      prog_detail <- "Calculate, Create Report"
+      message(paste0("\n", prog_detail))
+      # Increment the progress bar, and update the detail text.
+      incProgress(1/prog_n, detail = prog_detail)
+      Sys.sleep(2 * prog_sleep)
+      
+      strFile.RMD <- file.path("external", "Results_Summary.Rmd")
+      strFile.RMD.format <- "html_document"
+      strFile.out <- paste0(fn_input_base, "_modtherm_RESULTS.html")
+      dir.export <- path_results
+      rmarkdown::render(strFile.RMD
+                        , output_format = strFile.RMD.format
+                        , output_file = strFile.out
+                        , output_dir = dir.export
+                        , quiet = TRUE)
+      
+      # Calc, 9, Clean Up----
+      prog_detail <- "Calculate, Clean Up"
+      message(paste0("\n", prog_detail))
+      # Increment the progress bar, and update the detail text.
+      incProgress(1/prog_n, detail = prog_detail)
+      Sys.sleep(2 * prog_sleep)
+      
+      # Create zip file of results
+      fn_4zip <- list.files(path = path_results
+                            , full.names = TRUE)
+      utils::zip(file.path(path_results, "results.zip"), fn_4zip)
+      
+      # button, enable, download
+      shinyjs::enable("b_download_modtherm")
+      
+    }## expr ~ withProgress ~ END
+    , message = "Calculating Thermal Model"
+    )## withProgress ~ END
+  }##expr ~ ObserveEvent ~ END
+  )##observeEvent ~ b_calc_modtherm ~ END
+  
+  # b_download_modtherm ----
+  output$b_download_modtherm <- downloadHandler(
+    
+    filename = function() {
+      inFile <- input$fn_input
+      fn_input_base <- tools::file_path_sans_ext(inFile$name)
+      paste0(fn_input_base
+             , "_ModTherm_"
+             , format(Sys.time(), "%Y%m%d_%H%M%S")
+             , ".zip")
+    } ,
+    content = function(fname) {##content~START
+      
+      file.copy(file.path(path_results, "results.zip"), fname)
+      
+    }##content~END
+    #, contentType = "application/zip"
+  )##download ~ BCG
+  
   
   # MTTI ----